+++ source/main/include/cpath/private/c_memory.h
@@ -12,15 +12,15 @@ namespace ncore
         struct memory_t
         {
             u8* m_ptr;       // Virtual memory
-            u32 m_reserved;  // Maximum reserved item count (unit=item size), memory size = m_reserved * item size
-            u32 m_committed; // Current committed item count (unit=item size), memory size = m_committed * item size
+            u64 m_reserved;  // Maximum reserved item count (unit=item size), memory size = m_reserved * item size
+            u64 m_committed; // Current committed item count (unit=item size), memory size = m_committed * item size
 
             // Note: You need to stick to a particular item size when using this interface
 
-            void init(u32 initial_capacity, u32 max_capacity, u32 item_size = 1);
+            void init(u64 initial_capacity, u64 max_capacity, u32 item_size = 1);
             void exit();
-            void reset(u32 initial_capacity, u32 item_size = 1);
-            void add_capacity(u32 add_capacity, u32 item_size = 1);
+            void reset(u64 initial_capacity, u32 item_size = 1);
+            void add_capacity(u64 add_capacity, u32 item_size = 1);
 
             void ensure_capacity(u32 index, u32 item_size = 1, u32 add_capacity_when_needed = 64 * 1024);
 
@@ -40,12 +40,12 @@ namespace ncore
         {
             memory_t m_memory;
 
-            void init(u32 initial_capacity, u32 max_capacity) { m_memory.init(initial_capacity, max_capacity, sizeof(T)); }
+            void init(u64 initial_capacity, u64 max_capacity) { m_memory.init(initial_capacity, max_capacity, sizeof(T)); }
             void exit() { m_memory.exit(); }
-            void reset(u32 initial_capacity) { m_memory.reset(initial_capacity, sizeof(T)); }
-            void add_capacity(u32 add_capacity) { m_memory.add_capacity(add_capacity, sizeof(T)); }
+            void reset(u64 initial_capacity) { m_memory.reset(initial_capacity, sizeof(T)); }
+            void add_capacity(u64 add_capacity) { m_memory.add_capacity(add_capacity, sizeof(T)); }
 
-            void ensure_capacity(u32 capacity, u32 add_capacity_when_needed = 1024 * 1024) { m_memory.ensure_capacity(capacity, add_capacity_when_needed); }
+            void ensure_capacity(u64 capacity, u32 add_capacity_when_needed = 1024 * 1024) { m_memory.ensure_capacity(capacity, add_capacity_when_needed); }
 
             T* allocate(u32 items) { return (T*)m_memory.allocate((u8*&)m_memory.m_ptr, items, sizeof(T)); }
             T* reserve(u32 items) { return (T*)m_memory.reserve(m_memory.m_ptr, items, sizeof(T)); }
